+++ src/work/future/ext.rs
@@ -1,4 +1,4 @@
-// devela::work::async::ext
+// devela::work::future::ext
 //
 //!
 //