+++ src/work/future/mod.rs
@@ -1,9 +1,10 @@
-// devela::work::async
+// devela::work::future
 //
 //! Asynchrony.
+// #![doc = crate::doc_!(modules: crate::work; future)]
+// #![doc = crate::doc_!(newline)]
+//!
 #![doc = crate::doc_!(extends: future, task)]
-#![doc = crate::doc_!(modules: crate::work; async)]
-#![doc = crate::doc_!(newline)]
 //!
 //! See also the fundamental [`async`] and [`await`] keywords and the
 //! [async book](https://rust-lang.github.io/async-book/).