+++ src/work/mod.rs
@@ -1,7 +1,7 @@
 // devela::work
 //
 //! Work management, concurrency handling.
-#![doc = crate::doc_!(modules: crate; work)]
+#![doc = crate::doc_!(modules: crate; work: future, sync, thread)]
 #![doc = crate::doc_!(newline)]
 //!
 #![doc = crate::doc_!(extends: future, process, sync, task, thread)]
@@ -9,27 +9,24 @@
 // safety
 #![cfg_attr(feature = "safe_work", forbid(unsafe_code))]
 
-mod r#async;
-mod sync;
-mod thread;
+pub mod future;
+pub mod sync;
+pub mod thread;
 
-crate::items! { // structural access: _mods, _all, _always
-    #[allow(unused)]
-    pub use _mods::*;
+crate::items! { // structural access: _pub_mods, _all, _always
     #[allow(unused)] #[doc(hidden, no_inline)]
-    pub use _always::*;
+    pub use {_always::*, _pub_mods::*};
 
-    mod _mods {
-        pub use super::r#async::_all::*;
-
+    mod _pub_mods { #![allow(unused)]
+        pub use super::future::_all::*;
         #[allow(unused, reason = "feature-gated")]
         pub use super::{sync::_all::*, thread::_all::*};
     }
     pub(super) mod _all {
         #[doc(inline)]
-        pub use super::_mods::*;
+        pub use super::_pub_mods::*;
     }
     pub(super) mod _always { #![allow(unused)]
-        pub use super::{r#async::_always::*, sync::_always::*, thread::_always::*};
+        pub use super::{future::_always::*, sync::_always::*, thread::_always::*};
     }
 }