+++ DOCS/CHANGELOG.md
@@ -26,9 +26,10 @@
   - `_docs_max` to `_max`, `_docs_min` to `_docs`.
 - rename re-exports: `Layout` to `MemLayout`, `LayoutError` to `MemLayoutError`.
 - rename `LoggerConfig` to `LogConfig`.
+- rename `work::async` to `work::future`.
 - ungate: `FxHasher`, `Xorshift128p`.
 - make customizable: `XorShift[16|32|64]`.
-- make public: `sys::env`.
+- make public: `sys::env`, `work::{future, sync, thread}`.
 
 ### Fixed
 - enable nightly features depending on `alloc` and `std`.